% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popprxl-package.r
\docType{package}
\name{popprxl}
\alias{popprxl}
\title{popprxl.}
\description{
This package has one function: \code{\link{read.genalexcel}}.
}
\section{Citation}{

To cite this package, please use the citation for the poppr package. You can
find it by typing \code{citation(package = "poppr")}.
}

\author{
Zhian N. Kamvar and Niklaus J. Grünwald
}
